Simplify launch files
