--- beluga_demo_nav2/launch/localization_launch.py
@@ -1,107 +1,0 @@
-# Copyright 2024 Ekumen, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import os
-
-from ament_index_python.packages import get_package_share_directory
-
-from launch import LaunchDescription
-from launch.actions import DeclareLaunchArgument
-from launch.substitutions import LaunchConfiguration
-from launch_ros.actions import LoadComposableNodes
-from launch_ros.descriptions import ComposableNode, ParameterFile
-from nav2_common.launch import RewrittenYaml
-
-
-def generate_launch_description():
-    # Get the launch directory
-    bringup_dir = get_package_share_directory('beluga_demo_nav2')
-
-    params_file = LaunchConfiguration('params_file')
-    container_name = LaunchConfiguration('container_name')
-
-    lifecycle_nodes = ['map_server', 'amcl']
-
-    # Create our own temporary YAML files that include substitutions
-    param_substitutions = {
-        'use_sim_time': 'True',
-        'yaml_filename': os.path.join(
-            bringup_dir, 'maps', 'turtlebot3_world.yaml'
-        ),  # noqa: E501
-    }
-
-    configured_params = ParameterFile(
-        RewrittenYaml(
-            source_file=params_file,
-            param_rewrites=param_substitutions,
-            convert_types=True,
-        ),
-        allow_substs=True,
-    )
-
-    declare_params_file_cmd = DeclareLaunchArgument(
-        'params_file',
-        default_value=os.path.join(bringup_dir, 'params', 'nav2_params.yaml'),
-        description='Full path to the ROS2 parameters'
-        'file to use for all launched nodes',
-    )
-
-    declare_container_name_cmd = DeclareLaunchArgument(
-        'container_name',
-        default_value='nav2_container',
-        description='the name of container that'
-        'nodes will load in if use composition',
-    )
-
-    load_composable_nodes = LoadComposableNodes(
-        target_container=container_name,
-        composable_node_descriptions=[
-            ComposableNode(
-                package='nav2_map_server',
-                plugin='nav2_map_server::MapServer',
-                name='map_server',
-                parameters=[configured_params],
-            ),
-            ComposableNode(
-                package='beluga_amcl',
-                plugin='beluga_amcl::AmclNode',
-                name='amcl',
-                parameters=[configured_params],
-            ),
-            ComposableNode(
-                package='nav2_lifecycle_manager',
-                plugin='nav2_lifecycle_manager::LifecycleManager',
-                name='lifecycle_manager_localization',
-                parameters=[
-                    {
-                        'use_sim_time': True,
-                        'autostart': True,
-                        'node_names': lifecycle_nodes,
-                    }
-                ],
-            ),
-        ],
-    )
-
-    # Create the launch description and populate
-    ld = LaunchDescription()
-
-    # Declare the launch options
-    ld.add_action(declare_params_file_cmd)
-    ld.add_action(declare_container_name_cmd)
-
-    # Add the actions to launch all of the localization nodes
-    ld.add_action(load_composable_nodes)
-
-    return ld
